# Use optimized distance calculation