@badisa 2021-05-04
Latest changes shows a large impact on performance, with almost a 40% improvement on `k_find_blocks_with_ixns`

```
==114274== Profiling application: /home/ubuntu/.conda/envs/timemachine/bin/python /home/ubuntu/.conda/envs/timemachine/bin/pytest tests/test_nblist.py -s
==114274== Profiling result:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   78.16%  4.8675ms        12  405.63us  17.952us  786.17us  void k_find_blocks_with_ixns<double>(int, double const *, double const *, double const *, double const *, unsigned int*, int*, unsigned int*, unsigned int*, double)
                    9.68%  602.59us        15  40.172us  39.648us  40.928us  k_find_block_bounds(int, int, int, double const *, double const *, double*, double*)
                    5.23%  325.60us        30  10.853us  1.0560us  64.128us  [CUDA memcpy HtoD]
                    5.15%  320.80us        42  7.6380us  1.2480us  79.968us  [CUDA memcpy DtoH]
                    1.06%  65.888us        12  5.4900us  5.0880us  5.9200us  k_compact_trim_atoms(int, int, unsigned int*, unsigned int*, int*, unsigned int*)
                    0.72%  44.960us        42  1.0700us  1.0240us  1.3760us  [CUDA memset]
      API calls:   95.97%  210.83ms        84  2.5099ms  2.1400us  210.32ms  cudaMalloc
                    2.47%  5.4251ms        15  361.67us  40.779us  823.64us  cudaDeviceSynchronize
                    0.92%  2.0260ms        72  28.139us  5.1310us  192.08us  cudaMemcpy
                    0.16%  352.10us        39  9.0280us  4.2930us  20.948us  cudaLaunchKernel
                    0.16%  351.24us         1  351.24us  351.24us  351.24us  cuDeviceTotalMem
                    0.15%  335.04us        78  4.2950us  2.2650us  18.077us  cudaFree
                    0.08%  172.14us       101  1.7040us     120ns  74.339us  cuDeviceGetAttribute
                    0.06%  140.93us        42  3.3550us  2.0820us  5.7310us  cudaMemsetAsync
                    0.01%  26.884us         1  26.884us  26.884us  26.884us  cuDeviceGetName
                    0.00%  8.7860us        39     225ns     164ns     367ns  cudaPeekAtLastError
                    0.00%  5.5860us         1  5.5860us  5.5860us  5.5860us  cuDeviceGetPCIBusId
                    0.00%  1.4160us         3     472ns     152ns  1.0530us  cuDeviceGetCount
                    0.00%     775ns         2     387ns     168ns     607ns  cuDeviceGet
                    0.00%     261ns         1     261ns     261ns     261ns  cuDeviceGetUuid
```